+++ docs/src/examples/components/Dropdown/Types/DropdownExampleSingleSearch.shorthand.tsx
@@ -13,19 +13,13 @@ const inputItems = [
   'Selina Kyle',
 ]
 
-const DropdownExampleMultipleSearch = () => (
+const DropdownExampleSingleSearch = () => (
   <Dropdown
     search
-    multiple
     items={inputItems}
     placeholder="Start typing a name"
-    getA11ySelectionMessage={getA11ySelectionMessage}
     noResultsMessage="We couldn't find any matches."
   />
 )
 
-const getA11ySelectionMessage = {
-  onAdd: item => `${item} has been selected.`,
-  onRemove: item => `${item} has been removed.`,
-}
-export default DropdownExampleMultipleSearch
+export default DropdownExampleSingleSearch