+++ packages/react/src/components/Dropdown/Dropdown.tsx
@@ -41,6 +41,8 @@ import ListItem from '../List/ListItem'
 
 export interface DropdownSlotClassNames {
   container: string
+  triggerButton: string
+  itemsList: string
   selectedItems: string
 }
 
@@ -218,11 +220,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
       }
 
       // targets DropdownItem shorthand objects
-      if ((item as any).header) {
-        return (item as any).header
-      }
-
-      return `${item}`
+      return (item as any).header || String(item)
     },
     toggleIndicator: {},
     triggerButton: {},
@@ -242,7 +240,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
       searchQuery: search ? '' : undefined,
       value: multiple ? [] : null,
       // used on single selection to open the dropdown with the selected option as highlighted.
-      defaultHighlightedIndex: !this.props.search && !this.props.multiple ? null : undefined,
+      defaultHighlightedIndex: this.props.multiple ? undefined : null,
     }
   }
 
@@ -265,7 +263,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
           itemToString={itemToString}
           // If it's single search, don't pass anything. Pass a null otherwise, as Downshift does
           // not handle selection by default for single/multiple selection and multiple search.
-          selectedItem={search && !multiple ? undefined : null}
+          selectedItem={null}
           getA11yStatusMessage={getA11yStatusMessage}
           defaultHighlightedIndex={defaultHighlightedIndex}
           onStateChange={this.handleStateChange}
@@ -289,7 +287,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
               <Ref innerRef={innerRef}>
                 <div
                   className={cx(Dropdown.slotClassNames.container, classes.container)}
-                  onClick={multiple ? this.handleContainerClick.bind(this, isOpen) : undefined}
+                  onClick={search && !isOpen ? this.handleContainerClick : undefined}
                 >
                   <div
                     ref={this.selectedItemsRef}
@@ -337,13 +335,13 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
     styles: ComponentSlotStylesInput,
     getToggleButtonProps: (options?: GetToggleButtonPropsOptions) => any,
   ): JSX.Element {
-    const { placeholder, itemToString, multiple, triggerButton } = this.props
-    const { value } = this.state
-    const content = value && !multiple ? itemToString(value) : placeholder
+    const content = this.getSelectedItemAsString(this.state.value)
+
     return (
       <Ref innerRef={this.buttonRef}>
-        {Button.create(triggerButton, {
+        {Button.create(this.props.triggerButton, {
           defaultProps: {
+            className: Dropdown.slotClassNames.triggerButton,
             content,
             fluid: true,
             styles: styles.button,
@@ -438,6 +436,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
         }}
       >
         <List
+          className={Dropdown.slotClassNames.itemsList}
           {...accessibilityMenuProps}
           styles={styles.list}
           tabIndex={search ? undefined : -1} // needs to be focused when trigger button is activated.
@@ -556,11 +555,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
   private getItemsFilteredBySearchQuery = (): ShorthandValue[] => {
     const { items, itemToString, multiple, search } = this.props
     const { searchQuery, value } = this.state
-    let filteredItems = items
-
-    if (multiple) {
-      filteredItems = _.difference(filteredItems, value as ShorthandValue[])
-    }
+    const filteredItems = multiple ? _.difference(items, value as ShorthandValue[]) : items
 
     if (search) {
       if (_.isFunction(search)) {
@@ -695,8 +690,8 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
     }
   }
 
-  private handleContainerClick = (isOpen: boolean) => {
-    !isOpen && this.inputRef.current.focus()
+  private handleContainerClick = () => {
+    this.tryFocusSearchInput()
   }
 
   private handleListKeyDown = (
@@ -716,7 +711,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
         return
       case keyboardKey.Escape:
         accessibilityInputPropsKeyDown(e)
-        this.buttonRef.current.focus()
+        this.tryFocusTriggerButton()
         return
       default:
         accessibilityInputPropsKeyDown(e)
@@ -725,12 +720,15 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
   }
 
   private handleSelectedChange = (item: ShorthandValue) => {
-    const { items, multiple, getA11ySelectionMessage, search } = this.props
-    const newValue = multiple ? [...(this.state.value as ShorthandValue[]), item] : item
+    const { items, multiple, getA11ySelectionMessage } = this.props
+    const newState = {
+      value: multiple ? [...(this.state.value as ShorthandValue[]), item] : item,
+      searchQuery: this.getSelectedItemAsString(item),
+    }
 
-    this.trySetState({ value: newValue, searchQuery: '' })
+    this.trySetState(newState)
 
-    if (!search && !multiple) {
+    if (!multiple) {
       this.setState({ defaultHighlightedIndex: items.indexOf(item) })
     }
 
@@ -746,22 +744,16 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
       )
     }
 
-    if (!search) {
-      this.buttonRef.current.focus()
-    }
+    this.tryFocusTriggerButton()
 
     // we don't have event for it, but want to keep the event handling interface, event is empty.
-    _.invoke(
-      this.props,
-      'onSelectedChange',
-      {},
-      { ...this.props, searchQuery: '', value: newValue },
-    )
+    _.invoke(this.props, 'onSelectedChange', {}, { ...this.props, ...newState })
   }
 
   private handleSelectedItemRemove(e: React.SyntheticEvent, item: ShorthandValue) {
     this.removeItemFromValue(item)
-    this.inputRef.current.focus()
+    this.tryFocusSearchInput()
+    this.tryFocusTriggerButton()
     e.stopPropagation()
   }
 
@@ -785,10 +777,44 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
     // we don't have event for it, but want to keep the event handling interface, event is empty.
     _.invoke(this.props, 'onSelectedChange', {}, { ...this.props, value })
   }
+
+  private tryFocusTriggerButton = () => {
+    if (!this.props.search) {
+      this.buttonRef.current.focus()
+    }
+  }
+
+  private tryFocusSearchInput = () => {
+    if (this.props.search) {
+      this.inputRef.current.focus()
+    }
+  }
+
+  /**
+   * if there is no value we use the placeholder value
+   * otherwise, for single selection we convert the value with itemToString
+   * and for multiple selection we return empty string, the values are rendered by renderSelectedItems
+   */
+  private getSelectedItemAsString = (value: ShorthandValue): string => {
+    const { itemToString, multiple, placeholder } = this.props
+    const isValueEmpty = _.isArray(value) ? value.length < 1 : !value
+
+    if (isValueEmpty) {
+      return placeholder
+    }
+
+    if (multiple) {
+      return ''
+    }
+
+    return itemToString(value)
+  }
 }
 
 Dropdown.slotClassNames = {
   container: `${Dropdown.className}__container`,
+  triggerButton: `${Dropdown.className}__trigger-button`,
+  itemsList: `${Dropdown.className}__items-list`,
   selectedItems: `${Dropdown.className}__selected-items`,
 }